use smaller font for code blocks to fit better in the layout
code blocks do not wrap at all, and if they are wide, then the sidebar will
be pushed over to the right, sometimes quite far.  This just reduces the
font size a little bit, bringing it inline with the regular text font size.
And it fixes most of the wideness issues that I've seen.  But there is
probably a better way to do this.