Move download button above screenshot.
Although I appreciate that we are trying to port the existing site as
close as possible before making decisions about what to change, I
couldn't go passed this one. On almost every non-giant screen, the
download button will not be visible when a user first loads the page. By
moving it above the screenshot, it will almost always be visible to all
users without the need to scroll.