Don't underline app list items when hovering.
It looked a little off when everything (title/summary/license) all get
underlined, no matter what is hovered - even if it was the icon.

However, I like that you can click anywhere to view the app, so this
just removes the underline but leaves the whole thing as a link.