Reduce space taken by app list in sidebar.
Also adjusts how the license is shown in the main list of apps.
Instead of adding a left margin, which causes some interesting wrapping
behaviour when applied to an inline span, this puts the license in
brackets and removes the italics. The brackets are added via CSS for
now, but they could equally be added in the template.