Don't show all possible pages when paginating. Applied styles to page buttons.
This only shows 4 pages either side of the current page, and optionally
the first or last page, depending on whether that is included in the "4
pages either side" or not. Also shows a previous and next button as per
before.

It no longer says "Page X of Y", because you can see that from the
paginator. It always has the first and the last page shown.