Fix redirect from old to new Atom feed
When redirecting with a HTML file from a XML file a XML Parsing Error is
thrown in many browsers.

By putting a slash at the end of the file name we trick out this
mechanism as the web server will try to search for a directory if a file
is not found and will then serve the index.html which will redirect to
the new file.